test: simplify
